+++ packages/design-system/ui/src/components/hint/hint.tsx
@@ -1,29 +1,35 @@
-import { VariantProps, cva } from "class-variance-authority"
+import { VariantProps, cva } from "cva"
 import * as React from "react"
 
 import { ExclamationCircleSolid } from "@medusajs/icons"
 import { clx } from "../../utils/clx"
 
-const hintVariants = cva(
-  "txt-compact-xsmall inline-flex items-center gap-x-2",
-  {
-    variants: {
-      variant: {
-        info: "text-ui-fg-subtle",
-        error: "text-ui-fg-error",
-      },
+const hintVariants = cva({
+  base: "txt-compact-xsmall inline-flex items-center gap-x-2",
+  variants: {
+    variant: {
+      info: "text-ui-fg-subtle",
+      error: "text-ui-fg-error",
     },
-    defaultVariants: {
-      variant: "info",
-    },
-  }
-)
+  },
+  defaultVariants: {
+    variant: "info",
+  },
+})
 
-type HintProps = VariantProps<typeof hintVariants> &
-  React.ComponentPropsWithoutRef<"span">
+interface HintProps extends VariantProps<typeof hintVariants>,
+  React.ComponentPropsWithoutRef<"span"> {}
 
 const Hint = React.forwardRef<HTMLSpanElement, HintProps>(
-  ({ className, variant = "info", children, ...props }, ref) => {
+  ({ 
+    className, 
+    /**
+     * The hint's style.
+     */
+    variant = "info", 
+    children, 
+    ...props 
+  }: HintProps, ref) => {
     return (
       <span
         ref={ref}