+++ packages/medusa/src/loaders/plugins.ts
@@ -1,4 +1,8 @@
-import { promiseAll, SearchUtils, upperCaseFirst } from "@medusajs/utils"
+import {
+  AbstractSearchService,
+  promiseAll,
+  upperCaseFirst,
+} from "@medusajs/utils"
 import { aliasTo, asFunction, asValue, Lifetime } from "awilix"
 import { Express } from "express"
 import fs from "fs"
@@ -6,19 +10,19 @@ import { sync as existsSync } from "fs-exists-cached"
 import glob from "glob"
 import _ from "lodash"
 import { createRequireFromPath } from "medusa-core-utils"
-import { FileService, OauthService } from "medusa-interfaces"
+import { OauthService } from "medusa-interfaces"
 import { trackInstallation } from "medusa-telemetry"
 import { EOL } from "os"
 import path from "path"
 import { EntitySchema } from "typeorm"
 import {
+  AbstractBatchJobStrategy,
+  AbstractCartCompletionStrategy,
+  AbstractFileService,
+  AbstractNotificationService,
+  AbstractPriceSelectionStrategy,
+  AbstractTaxCalculationStrategy,
   AbstractTaxService,
-  isBatchJobStrategy,
-  isCartCompletionStrategy,
-  isFileService,
-  isNotificationService,
-  isPriceSelectionStrategy,
-  isTaxCalculationStrategy,
 } from "../interfaces"
 import { MiddlewareService } from "../services"
 import {
@@ -35,9 +39,7 @@ import { getModelExtensionsMap } from "./helpers/get-model-extension-map"
 import ScheduledJobsLoader from "./helpers/jobs"
 import {
   registerAbstractFulfillmentServiceFromClass,
-  registerFulfillmentServiceFromClass,
   registerPaymentProcessorFromClass,
-  registerPaymentServiceFromClass,
 } from "./helpers/plugins"
 import { RoutesLoader } from "./helpers/routing"
 import { SubscriberLoader } from "./helpers/subscribers"
@@ -230,7 +232,9 @@ export function registerStrategies(
     const module = require(file).default
 
     switch (true) {
-      case isTaxCalculationStrategy(module.prototype): {
+      case AbstractTaxCalculationStrategy.isTaxCalculationStrategy(
+        module.prototype
+      ): {
         if (!("taxCalculationStrategy" in registeredServices)) {
           container.register({
             taxCalculationStrategy: asFunction(
@@ -246,7 +250,9 @@ export function registerStrategies(
         break
       }
 
-      case isCartCompletionStrategy(module.prototype): {
+      case AbstractCartCompletionStrategy.isCartCompletionStrategy(
+        module.prototype
+      ): {
         if (!("cartCompletionStrategy" in registeredServices)) {
           container.register({
             cartCompletionStrategy: asFunction(
@@ -262,7 +268,7 @@ export function registerStrategies(
         break
       }
 
-      case isBatchJobStrategy(module.prototype): {
+      case AbstractBatchJobStrategy.isBatchJobStrategy(module.prototype): {
         container.registerAdd(
           "batchJobStrategies",
           asFunction((cradle) => new module(cradle, pluginDetails.options))
@@ -279,7 +285,9 @@ export function registerStrategies(
         break
       }
 
-      case isPriceSelectionStrategy(module.prototype): {
+      case AbstractPriceSelectionStrategy.isPriceSelectionStrategy(
+        module.prototype
+      ): {
         if (!("priceSelectionStrategy" in registeredServices)) {
           container.register({
             priceSelectionStrategy: asFunction(
@@ -441,8 +449,7 @@ async function registerApi(
 }
 
 /**
- * Registers a service at the right location in our container. If the service is
- * a BaseService instance it will be available directly from the container.
+ * Registers a service at the right location in our container.
  * PaymentService instances are added to the paymentProviders array in the
  * container. Names are camelCase formatted and namespaced by the folder i.e:
  * services/example-payments -> examplePaymentsService
@@ -464,13 +471,10 @@ export async function registerServices(
 
       const context = { container, pluginDetails, registrationName: name }
 
-      registerPaymentServiceFromClass(loaded, context)
       registerPaymentProcessorFromClass(loaded, context)
-
-      registerFulfillmentServiceFromClass(loaded, context)
       registerAbstractFulfillmentServiceFromClass(loaded, context)
 
-      if (loaded.prototype instanceof OauthService) {
+      if (OauthService.isOauthService(loaded.prototype)) {
         const appDetails = loaded.getAppDetails(pluginDetails.options)
 
         const oauthService =
@@ -486,7 +490,9 @@ export async function registerServices(
             }
           ),
         })
-      } else if (isNotificationService(loaded.prototype)) {
+      } else if (
+        AbstractNotificationService.isNotificationService(loaded.prototype)
+      ) {
         container.registerAdd(
           "notificationProviders",
           asFunction((cradle) => new loaded(cradle, pluginDetails.options), {
@@ -505,10 +511,7 @@ export async function registerServices(
           ),
           [`noti_${loaded.identifier}`]: aliasTo(name),
         })
-      } else if (
-        loaded.prototype instanceof FileService ||
-        isFileService(loaded.prototype)
-      ) {
+      } else if (AbstractFileService.isFileService(loaded.prototype)) {
         // Add the service directly to the container in order to make simple
         // resolution if we already know which file storage provider we need to use
         container.register({
@@ -520,7 +523,7 @@ export async function registerServices(
           ),
           [`fileService`]: aliasTo(name),
         })
-      } else if (SearchUtils.isSearchService(loaded.prototype)) {
+      } else if (AbstractSearchService.isSearchService(loaded.prototype)) {
         // Add the service directly to the container in order to make simple
         // resolution if we already know which search provider we need to use
         container.register({
@@ -534,7 +537,7 @@ export async function registerServices(
         })
 
         container.register(isSearchEngineInstalledResolutionKey, asValue(true))
-      } else if (loaded.prototype instanceof AbstractTaxService) {
+      } else if (AbstractTaxService.isTaxService(loaded.prototype)) {
         container.registerAdd(
           "taxProviders",
           asFunction((cradle) => new loaded(cradle, pluginDetails.options), {