+++ packages/utils/src/promotion/index.ts
@@ -1,0 +1,20 @@
+export enum PromotionType {
+  STANDARD = "standard",
+  BUYGET = "buyget",
+}
+
+export enum ApplicationMethodType {
+  FIXED = "fixed",
+  PERCENTAGE = "percentage",
+}
+
+export enum ApplicationMethodTargetType {
+  ORDER = "order",
+  SHIPPING = "shipping",
+  ITEM = "item",
+}
+
+export enum ApplicationMethodAllocation {
+  EACH = "each",
+  ACROSS = "across",
+}